+++ package.json
@@ -14,6 +14,7 @@
     "remove": "node ./src/fs/delete",
     "list": "node ./src/fs/list",
     "read": "node ./src/fs/read",
+    "cjsToEsm": "node ./src/modules/cjsToEsm.mjs",
     "env": "RSS_USER_NAME='user name' RSS_USER_SURNAME='user surname' node ./src/cli/env set ",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
