+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,36 @@
+import path from "path";
+import {fileURLToPath} from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.resolve();
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import * as aJson from "./files/a.json" assert { type: "json" };
+import * as bJson from "./files/b.json" assert { type: "json" };
+import "./files/c.js";
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = aJson;
+} else {
+  unknownObject = bJson;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+export default {
+  unknownObject,
+    createMyServer,
+};
